feat: Changed the file hierarchy for the entire project (#3)
