+++ src/components/sections/FeaturedPostsSection.jsx
