+++ src/components/sections/FeaturedProductsSection.jsx
