--- src/layout/DefaultLayout.jsx
@@ -1,12 +1,0 @@
-import TopBar from "../components/layout/TopBar";
-
-function DefaultLayout({ children }) {
-  return (
-    <>
-      <TopBar />
-      {/*Navbar*/}
-      <main>{children}</main>
-      {/*Footer*/}
-    </>
-  );
-}
