+++ src/layouts/DefaultLayout.jsx
@@ -1,0 +1,16 @@
+import Footer from "./Footer";
+import Navbar from "./Navbar";
+import TopBar from "./TopBar";
+
+function DefaultLayout({ children }) {
+  return (
+    <>
+      <TopBar />
+      <Navbar />
+      <main>{children}</main>
+      <Footer />
+    </>
+  );
+}
+
+export default DefaultLayout;
